face-recognition-with-speech/face/requirements.txt
numpy>=1.24.0,<2.1
torch==2.9.1
Pillow>=8.0.0
facenet-pytorch>=2.3.0
supabase==2.5.0
python-dotenv>=1.0.0
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
requests>=2.28.0
pyttsx3>=2.90